Move grpc agent to its "final" dentination
Mostly just moving the test grpc agent from #75 to a place that matches the HTTP agent placement.

Bonus:

- Call go build to both client & server, so we can actually easily bulid & manually test them.
- Fix a bug with `RRB_IGNORE_PATTERN` not ignoring generated `.go` files, so RRB would just re-trigger itself and never finish.

commit-id:b4b6383b
