+++ internal/host/agent_server_grpc/client/main.go
@@ -8,7 +8,7 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 
-	"github.com/fornellas/resonance/internal/grpc/proto"
+	"github.com/fornellas/resonance/internal/host/agent_server_grpc/proto"
 )
 
 func main() {
